feat(item-details): open the image at large in modal
